to do

+++ Source/RenderEngine/UserInterface/HumanView.h
@@ -16,7 +16,7 @@ public:
 	virtual HRESULT VOnRestore();
 	virtual HRESULT VOnLostDevice();
 	virtual void VOnUpdate(uint32_t deltaMilliseconds);
-	virtual void VOnRender(float fTime, float fElapsedTime);
+	virtual void VOnRender(double fTime, float fElapsedTime);
 	virtual GameViewType VGetType() { return GameView_Human; }
 	virtual GameViewId VGetId() const { return m_ViewId; }
